+++ src/main/java/wooteco/subway/common/BaseControllerAdvice.java
@@ -1,4 +1,4 @@
-package wooteco.subway;
+package wooteco.subway.common;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
