+++ src/main/java/wooteco/subway/common/ExceptionMessageObj.java
@@ -1,4 +1,4 @@
-package wooteco.subway;
+package wooteco.subway.common;
 
 public class ExceptionMessageObj {
     private String errorMessage;
