+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -11,7 +11,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.ExceptionMessageObj;
 
 import java.util.HashMap;
 import java.util.List;
